requirements.txt
Flask
flask-cors
requests
discogs_client
python-dotenv
gunicorn
